+++ libbz2-rs-sys/src/huffman.rs
@@ -179,8 +179,8 @@ pub(crate) fn create_decode_tables(
     base: &mut [i32],
     perm: &mut [i32],
     length: &mut [u8],
-    minLen: i32,
-    maxLen: i32,
+    minLen: u8,
+    maxLen: u8,
     alphaSize: i32,
 ) {
     let alphaSize = usize::try_from(alphaSize).unwrap_or(0);
@@ -188,7 +188,7 @@ pub(crate) fn create_decode_tables(
     let mut pp: i32 = 0;
     for i in minLen..=maxLen {
         for (j, e) in length[0..alphaSize].iter().enumerate() {
-            if *e as i32 == i {
+            if *e == i {
                 perm[pp as usize] = j as i32;
                 pp += 1;
             }
